mat: add Einsum function
This commit adds an Einsum function that emulates a subset of the
behavior of the numpy.einsum function, and equivalent ones in other
libraries like pytorch. In doing so, we allow for einstein summation
notation shorthand to be used to execute a variety of operations on
input matrixes.

In keeping with what appear to be the semantics of the package, error
handling is performed through panics, and it is up to the caller to make
sure that their inputs are fully correct.

Numpy docs for the equivalent function, which inspired terminology here:
- https://numpy.org/doc/stable/reference/generated/numpy.einsum.html

Understanding of einsum and test cases based on:
- https://ajcr.net/Basic-guide-to-einsum/
- https://www.youtube.com/watch?v=pkVwUVEHmfI